Refactor user related code
