+++ core/src/main/kotlin/no/ntnu/beardblaster/screen/LoginMenuScreen.kt
@@ -21,6 +21,8 @@ import no.ntnu.beardblaster.assets.Nls
 import no.ntnu.beardblaster.commons.State
 import no.ntnu.beardblaster.commons.user.User
 import no.ntnu.beardblaster.ui.*
+import no.ntnu.beardblaster.user.LoginDialog
+import no.ntnu.beardblaster.user.RegisterDialog
 import no.ntnu.beardblaster.user.UserAuth
 import no.ntnu.beardblaster.user.UserRepository
 import pl.mk5.gdx.fireapp.auth.GdxFirebaseUser
